ChatApp/requirements.txt
absl-py==0.6.1
asgiref==3.2.3
astor==0.7.1
certifi==2019.11.28
chardet==3.0.4
CodeConvert==3.0.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-cors-headers==3.2.0
django-models-ext==1.1.9
django-rest-swagger==2.2.0
django-six==1.0.4
djangorestframework==3.11.0
furl==2.1.0
gast==0.2.0
grpcio==1.16.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.1
numpy==1.18.1
openapi-codec==1.3.2
orderedmultidict==1.0.1
protobuf==3.6.1
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
shortuuid==0.5.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
termcolor==1.1.0
TimeConvert==1.5.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
wincertstore==0.2
xlwt==1.3.0
